Vaccine Management System
In this assignment, you implement several C++ classes for a vaccine management system. This
assignment aims to create a storage system for the vaccines. It will allow the users to search for details
of available vaccines and their total cost to the related vaccination center (hospital hubs, local vaccination services, pharmacies, etc.). Put each of the class declarations in a separate header file and implement the member methods in separate .cpp files.